RUM-4098: Expose datastore as a property in corefeature
