+++ dd-sdk-android-core/api/apiSurface
@@ -102,10 +102,8 @@ interface com.datadog.android.api.feature.FeatureContextUpdateReceiver
 interface com.datadog.android.api.feature.FeatureEventReceiver
   fun onReceive(Any)
 interface com.datadog.android.api.feature.FeatureScope
+  var dataStore: com.datadog.android.core.internal.persistence.datastore.DataStoreHandler
   fun withWriteContext(Boolean = false, (com.datadog.android.api.context.DatadogContext, com.datadog.android.api.storage.EventBatchWriter) -> Unit)
-  fun <T: Any> writeToDataStore(String, String, com.datadog.android.core.persistence.Serializer<T>, T)
-  fun <T: Any> readFromDataStore(String, String, com.datadog.android.core.internal.persistence.Deserializer<String, T>, Int): T?
-  fun getDataStoreCurrentVersion(): Int
   fun sendEvent(Any)
   fun <T: Feature> unwrap(): T
 fun <R: Any?> com.datadog.android.api.InternalLogger.measureMethodCallPerf(Class<*>, String, Float = 100f, () -> R): R
@@ -249,6 +247,11 @@ interface com.datadog.android.core.internal.net.FirstPartyHostHeaderTypeResolver
   fun isEmpty(): Boolean
 interface com.datadog.android.core.internal.persistence.Deserializer<P: Any, R: Any>
   fun deserialize(P): R?
+interface com.datadog.android.core.internal.persistence.datastore.DataStoreHandler
+  fun <T: Any> write(String, String, com.datadog.android.core.persistence.Serializer<T>, T)
+  fun <T: Any> read(String, String, com.datadog.android.core.internal.persistence.Deserializer<String, T>, Int): T?
+  companion object 
+    const val CURRENT_DATASTORE_VERSION: Int
 fun java.io.File.canReadSafe(com.datadog.android.api.InternalLogger): Boolean
 fun java.io.File.existsSafe(com.datadog.android.api.InternalLogger): Boolean
 fun java.io.File.readTextSafe(java.nio.charset.Charset = Charsets.UTF_8, com.datadog.android.api.InternalLogger): String?
