+++ dd-sdk-android-core/src/main/kotlin/com/datadog/android/core/internal/persistence/datastore/DataStoreHandler.kt
@@ -9,9 +9,17 @@ package com.datadog.android.core.internal.persistence.datastore
 import com.datadog.android.core.internal.persistence.Deserializer
 import com.datadog.android.core.persistence.Serializer
 
-internal const val CURRENT_DATASTORE_VERSION: Int = 0
+interface DataStoreHandler {
 
-internal interface DataStoreHandler {
+    /**
+     * Write data to the datastore.
+     *
+     * @param T datatype of the data to write to the datastore.
+     * @param dataStoreFileName name of the datastore file as there could be multiple such files per feature.
+     * @param featureName of the calling feature, to determine the path to the datastore file.
+     * @param serializer to use to serialize the data.
+     * @param data to write.
+     */
     fun <T : Any> write(
         dataStoreFileName: String,
         featureName: String,
@@ -19,10 +27,26 @@ internal interface DataStoreHandler {
         data: T
     )
 
+    /**
+     * Read data from the datastore.
+     *
+     * @param T datatype of the data to read from the datastore.
+     * @param dataStoreFileName name of the datastore file as there could be multiple such files per feature.
+     * @param featureName of the calling feature, to determine the path to the datastore file.
+     * @param deserializer to use to deserialize the data.
+     * @param version to use when reading from the datastore (to support migrations).
+     */
     fun <T : Any> read(
         dataStoreFileName: String,
         featureName: String,
         deserializer: Deserializer<String, T>,
         version: Int
     ): T?
+
+    companion object {
+        /**
+         * The current version of the datastore.
+         */
+        const val CURRENT_DATASTORE_VERSION: Int = 0
+    }
 }
